🚚 [MOVE] : 메세지 카드 Grid 컴포넌트 분리
