+++ src/components/MessageCard/MessageCard.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import React, { useRef, useState } from 'react';
 import * as S from './MessageCard.style';
 import { getFormattedDate, Deleted } from './index';
