+++ src/components/MessageCardHeader/MessageCardHeader.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import * as S from './MessageCardHeader.style';
 import { CardBadage } from '../CardBadage/CardBadage';
-
 export const MessageCardHeader = ({ cardData, Component }) => {
   return (
     <>
