+++ src/pages/PostIDPage/PostIDPage.style.js
@@ -1,4 +1,4 @@
-import { styled, keyframes } from 'styled-components';
+import { styled } from 'styled-components';
 import { COLORS } from '../../style/colorPalette';
 import { FONT_STYLE } from '../../style/fontStyle';
 
@@ -50,30 +50,6 @@ export const ModalBackground = styled.div`
   z-index: 1000;
 `;
 
-export const GridWrapper = styled.div`
-  display: grid;
-  width: 1200px;
-  grid-template-columns: 1fr 1fr 1fr;
-  gap: 24px;
-  padding: 113px 0;
-  margin: 0 auto;
-
-  @media (max-width: 1248px) {
-    width: auto;
-    padding: 113px 24px;
-  }
-  @media (max-width: 1200px) {
-    grid-template-columns: 1fr 1fr;
-    max-width: 1020px;
-  }
-
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-    max-width: 495px;
-    min-width: 360px;
-  }
-`;
-
 export const ErrorWrapper = styled.div`
   padding: 15% 0;
   width: 100vw;
@@ -87,23 +63,4 @@ export const ErrorTitle = styled.p`
 export const ErrorContent = styled.p`
   margin-top: 15px;
   ${FONT_STYLE.REGULAR_15};
-`;
-
-const loading = keyframes`
-  from{
-    transfrom: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-`;
-
-export const LoadingIcon = styled.img`
-  visibility: ${({ $endData }) => ($endData ? 'hidden' : 'visible')};
-  position: absolute;
-  ${({ $initialLoading }) => ($initialLoading ? 'top: 40vh' : 'bottom: 30px')};
-  left: calc(50% - 15px);
-  width: 30px;
-  height: 30px;
-  animation: ${loading} 0.9s infinite;
 `;
