+++ src/pages/PostIDPage/index.js
@@ -2,8 +2,8 @@ import { AddMessageCard } from '../../components/AddMessageCard/AddMessageCard';
 import { PostIDContext } from '../../context/PostIDContext';
 import { MessageCard } from '../../components/MessageCard/MessageCard';
 import { Modal } from '../../components/Modal/Modal';
-import loadingIcon from '../../assets/icon/loading.svg';
 import { getMessageCardData } from '../../API';
+import { MessageCardWrapper } from '../../components/MessageCardWrapper/MessageCardWrapper';
 
 export {
   AddMessageCard,
@@ -11,5 +11,5 @@ export {
   MessageCard,
   Modal,
   getMessageCardData,
-  loadingIcon,
+  MessageCardWrapper,
 };
